#1116 - Allow creating a new project via page parameter
- Pass "NEW" as project id to create a new project
- Move "cancel" and "delete" buttons from the detail tab to the top-level so they are visible irrespective of which tab is selected
