+++ webanno-ui-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/project/ProjectPage.html
@@ -19,14 +19,23 @@
 <html xmlns:wicket="http://wicket.apache.org">
 <body>
   <wicket:extend>
+    <div wicket:id="deleteProjectDialog"></div>
     <div class="flex-content flex-h-container flex-gutter">
       <div wicket:id="sidebar" class="flex-sidebar flex-v-container flex-gutter">
         <wicket:container wicket:id="projects" />
         <wicket:container wicket:id="importPanel" />
       </div>
       <div wicket:id ="tabContainer" class="flex-content flex-v-container flex-gutter">
-        <div class="page-header">
-          <h1><wicket:container wicket:id="projectName"/></h1>
+        <div class="page-header flex-h-container">
+          <div class="flex-content">
+            <h1><wicket:container wicket:id="projectName"/></h1>
+          </div>
+          <div>
+            <h1>
+              <input wicket:id="delete" type="button" class="btn btn-danger" wicket:message="value:delete" />
+              <input wicket:id="cancel" type="button" class="btn btn-default" wicket:message="value:cancel" />
+            </h1>
+          </div>
         </div>
         <div wicket:id="tabPanel" class="flex-content flex-v-container"></div>
       </div>
