+++ webanno-ui-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/project/ProjectPage.java
@@ -20,27 +20,35 @@ package de.tudarmstadt.ukp.clarin.webanno.ui.project;
 import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.PAGE_PARAM_PROJECT_ID;
 import static de.tudarmstadt.ukp.clarin.webanno.security.model.Role.ROLE_ADMIN;
 import static de.tudarmstadt.ukp.clarin.webanno.security.model.Role.ROLE_PROJECT_CREATOR;
+import static de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaBehavior.visibleWhen;
+import static org.apache.wicket.authroles.authorization.strategies.role.metadata.MetaDataRoleAuthorizationStrategy.authorize;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 import javax.persistence.NoResultException;
 
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.wicket.Component;
-import org.apache.wicket.authroles.authorization.strategies.role.metadata.MetaDataRoleAuthorizationStrategy;
+import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.extensions.ajax.markup.html.tabs.AjaxTabbedPanel;
 import org.apache.wicket.extensions.markup.html.tabs.AbstractTab;
 import org.apache.wicket.extensions.markup.html.tabs.ITab;
+import org.apache.wicket.feedback.IFeedback;
 import org.apache.wicket.markup.html.WebMarkupContainer;
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.markup.html.panel.Panel;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.Model;
 import org.apache.wicket.model.PropertyModel;
+import org.apache.wicket.model.StringResourceModel;
 import org.apache.wicket.request.mapper.parameter.PageParameters;
 import org.apache.wicket.spring.injection.annot.SpringBean;
 import org.apache.wicket.util.string.StringValue;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.wicketstuff.annotation.mount.MountPath;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.ProjectService;
@@ -51,6 +59,8 @@ import de.tudarmstadt.ukp.clarin.webanno.security.UserDao;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.User;
 import de.tudarmstadt.ukp.clarin.webanno.support.ApplicationContextProvider;
 import de.tudarmstadt.ukp.clarin.webanno.support.bootstrap.BootstrapAjaxTabbedPanel;
+import de.tudarmstadt.ukp.clarin.webanno.support.dialog.ChallengeResponseDialog;
+import de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaAjaxLink;
 import de.tudarmstadt.ukp.clarin.webanno.support.wicket.ModelChangedVisitor;
 import de.tudarmstadt.ukp.clarin.webanno.ui.core.page.ApplicationPageBase;
 import de.tudarmstadt.ukp.clarin.webanno.ui.core.settings.ProjectSettingsPanelFactory;
@@ -72,6 +82,10 @@ import de.tudarmstadt.ukp.clarin.webanno.ui.project.users.ProjectUsersPanel;
 public class ProjectPage
     extends ApplicationPageBase
 {
+    private static final Logger LOG = LoggerFactory.getLogger(ProjectPage.class);
+    
+    public static final String NEW_PROJECT_ID = "NEW";
+
     private static final long serialVersionUID = -2102136855109258306L;
 
     // private static final Logger LOG = LoggerFactory.getLogger(ProjectPage.class);
@@ -87,6 +101,9 @@ public class ProjectPage
     private ProjectImportPanel importProjectPanel;
 
     private IModel<Project> selectedProject;
+    private ChallengeResponseDialog deleteProjectDialog;
+    
+    private boolean preSelectedModelMode = false;
     
     public ProjectPage()
     {
@@ -101,26 +118,32 @@ public class ProjectPage
         
         commonInit();
        
+        preSelectedModelMode = true;
+        
         sidebar.setVisible(false);
         
-        User user = userRepository.getCurrentUser();
-        
-        // Get current project from parameters
+        // Fetch project parameter
         StringValue projectParameter = aPageParameters.get(PAGE_PARAM_PROJECT_ID);
-        Optional<Project> project = getProjectFromParameters(projectParameter);
-        
-        if (project.isPresent()) {
-            // Check access to project
-            if (project != null && !projectService.isAdmin(project.get(), user)) {
-                error("You have no permission to access project [" + project.get().getId() + "]");
+        // Check if we are asked to create a new project
+        if (projectParameter != null && NEW_PROJECT_ID.equals(projectParameter.toString())) {
+            selectedProject.setObject(new Project());
+        }
+        // Check if we are asked to open an existing project
+        else {
+            Optional<Project> project = getProjectFromParameters(projectParameter);
+            if (project.isPresent()) {
+                // Check access to project
+                if (!projectService.isAdmin(project.get(), userRepository.getCurrentUser())) {
+                    error("You have no permission to access project [" + project.get().getId() + "]");
+                    setResponsePage(getApplication().getHomePage());
+                }
+                
+                selectedProject.setObject(project.get());
+            }
+            else {
+                error("Project [" + projectParameter + "] does not exist");
                 setResponsePage(getApplication().getHomePage());
             }
-            
-            selectedProject.setObject(project.get());
-        }
-        else {
-            error("Project [" + projectParameter + "] does not exist");
-            setResponsePage(getApplication().getHomePage());
         }
     }
     
@@ -133,10 +156,17 @@ public class ProjectPage
         add(sidebar);
 
         tabContainer = new WebMarkupContainer("tabContainer");
-        tabContainer.setOutputMarkupId(true);
+        tabContainer.setOutputMarkupPlaceholderTag(true);
+        tabContainer.add(visibleWhen(() -> selectedProject.getObject() != null));
         add(tabContainer);
         
         tabContainer.add(new Label("projectName", PropertyModel.of(selectedProject, "name")));
+        
+        tabContainer.add(new LambdaAjaxLink("cancel", this::actionCancel));
+        
+        tabContainer.add(new LambdaAjaxLink("delete", this::actionDelete)
+                .onConfigure((_this) -> _this.setEnabled(selectedProject.getObject() != null
+                        && selectedProject.getObject().getId() != null)));
         
         tabPanel = new BootstrapAjaxTabbedPanel<ITab>("tabPanel", makeTabs()) {
             private static final long serialVersionUID = -7356420977522213071L;
@@ -167,10 +197,34 @@ public class ProjectPage
         });
         sidebar.add(projects);
 
+        IModel<String> projectNameModel = PropertyModel.of(selectedProject, "name");
+        add(deleteProjectDialog = new ChallengeResponseDialog("deleteProjectDialog",
+                new StringResourceModel("DeleteProjectDialog.title", this),
+                new StringResourceModel("DeleteProjectDialog.text", this)
+                        .setModel(selectedProject).setParameters(projectNameModel),
+                projectNameModel));
+        deleteProjectDialog.setConfirmAction((target) -> {
+            try {
+                projectService.removeProject(selectedProject.getObject());
+                if (preSelectedModelMode) {
+                    setResponsePage(getApplication().getHomePage());
+                }
+                else {
+                    selectedProject.setObject(null);
+                    target.add(getPage());
+                }
+            }
+            catch (IOException e) {
+                LOG.error("Unable to remove project :" + ExceptionUtils.getRootCauseMessage(e));
+                error("Unable to remove project " + ":" + ExceptionUtils.getRootCauseMessage(e));
+                target.addChildren(getPage(), IFeedback.class);
+            }
+        });
+        
         importProjectPanel = new ProjectImportPanel("importPanel", selectedProject);
         sidebar.add(importProjectPanel);
-        MetaDataRoleAuthorizationStrategy.authorize(importProjectPanel, Component.RENDER,
-                String.join(",", ROLE_ADMIN.name(), ROLE_PROJECT_CREATOR.name()));    
+        authorize(importProjectPanel, Component.RENDER,
+                String.join(",", ROLE_ADMIN.name(), ROLE_PROJECT_CREATOR.name()));
     }
 
     private List<ITab> makeTabs()
@@ -243,5 +297,23 @@ public class ProjectPage
         catch (NoResultException e) {
             return Optional.empty();
         }
+    }
+    
+    private void actionCancel(AjaxRequestTarget aTarget)
+    {
+        if (preSelectedModelMode) {
+            setResponsePage(getApplication().getHomePage());
+        }
+        else {
+            selectedProject.setObject(null);
+            
+            // Reload whole page because master panel also needs to be reloaded.
+            aTarget.add(getPage());
+        }
+    }
+
+    private void actionDelete(AjaxRequestTarget aTarget)
+    {
+        deleteProjectDialog.show(aTarget);
     }
 }
