+++ webanno-ui-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/project/detail/ProjectDetailPanel.html
@@ -19,8 +19,6 @@
 <html xmlns:wicket="http://wicket.apache.org">
 <body>
 <wicket:panel>
-  <div wicket:id="deleteProjectDialog"></div>
-  
   <form wicket:id="form" class="flex-content flex-v-container">
     <div class="flex-content scrolling panel-body">
       <div class="container-fluid">
@@ -82,9 +80,7 @@
     </div>
     <div class="panel-footer text-right">
       <input wicket:id="save" type="submit" class="btn btn-primary" wicket:message="value:save" />
-      <input wicket:id="delete" type="button" class="btn btn-danger" wicket:message="value:delete" />
-      <input wicket:id="cancel" type="button" class="btn btn-default" wicket:message="value:cancel" />
-    </div>            
+    </div>
   </form>
 </wicket:panel>
 </body>
